Create unified "relay" package and move cursor connection to it
